Imported upstream version '2.3.16' of 'upstream'

+++ smacc2_performance_tools/_smacc2_sm_template/include/sm_name/sm_name.hpp
@@ -27,10 +27,10 @@ struct State1;
 struct State2;
 
 //VARIABLES - shared between states (using "_<name>_"-syntax to make this obvious)
-static uint64_t _counter_ = 0;
+static unsigned int _counter_ = 0;
 rclcpp::Time _start_time_;
 
-uint64_t _sum_of_iterations_ = 0.0;
+unsigned int _sum_of_iterations_ = 0.0;
 double _sum_of_elapsed_time_ = 0.0;
 
 //--------------------------------------------------------------------
